refactor(headers): de-leak rapidjson and <filesystem> from the public/cimport surface (CoolProp-h1ar)
The public headers transitively pulled two heavy/third-party deps into
every consumer of AbstractState.h and into the Cython cimport surface,
though neither appears in any State/AbstractState API:

  - rapidjson: constants_header.pxd did `cdef extern from
    "CoolProp/Configuration.h"` solely to read the configuration_keys
    enum; Configuration.h pulls detail/rapidjson.h for its inline JSON.
  - <filesystem> (and thus C++17): AbstractState.h -> detail/CachedElement.h
    -> detail/tools.h -> detail/filepaths.h, whose only std::filesystem
    use was the write_bytes_atomic() parameter.

Split each leaked symbol into a lean header the surface can consume:

  - detail/configuration_keys.h: the CONFIGURATION_KEYS_ENUM X-macro +
    `enum configuration_keys`, no includes. Configuration.h includes it;
    generate_constants_module.py reads the X-macro from it and emits
    `cdef extern from "CoolProp/detail/configuration_keys.h"` -> the
    cimport surface no longer pulls rapidjson.
  - detail/atomic_write.h: the write_bytes_atomic() declaration +
    <filesystem>. detail/filepaths.h drops <filesystem>; the three
    callers include the new header.

Also make fluids/PhaseEnvelope.h self-sufficient for CoolPropDbl
(include detail/tools.h): it used the type but relied on transitive
inclusion via the old constants_header.pxd -> Configuration.h edge, so
the leaner extern would otherwise break the generated CoolProp.cpp
(Cython module) include order.  Verified: generated CoolProp.cpp
compiles clean; cimport surface has no rapidjson path; -H shows no
rapidjson/<filesystem> in the AbstractState.h tree. Build + [SBTL] green.
fmt left in place (separate C++-standard-floor decision).
